atlas-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
openai==1.34.0
python-dotenv==1.0.1
sqlmodel==0.0.16
google-generativeai